+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * 1.7.1
   - added group feature to replace the previous suite and step group
+  - removed unpackSourceMap option, it should be true always
 
 
 * 1.7.0